big refactor, dyn iterator
