+++ libs/eversol_stake_pool/src/stakedex_traits/mod.rs
@@ -1,0 +1,9 @@
+mod base;
+mod deposit_sol;
+mod deposit_stake;
+mod withdraw_stake;
+
+pub use base::*;
+pub use deposit_sol::*;
+pub use deposit_stake::*;
+pub use withdraw_stake::*;
