+++ libs/marinade/src/stakedex_traits/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod deposit_sol;
+mod deposit_stake;
+
+pub use base::*;
+pub use deposit_sol::*;
+pub use deposit_stake::*;
