+++ stakedex_sdk/tests/test_main.rs
@@ -4,6 +4,7 @@ use solana_account_decoder::UiAccountEncoding;
 use solana_client::{
     rpc_client::RpcClient,
     rpc_config::{RpcSimulateTransactionAccountsConfig, RpcSimulateTransactionConfig},
+    rpc_response::{Response, RpcSimulateTransactionResult},
 };
 use solana_sdk::{account::Account, pubkey::Pubkey, signer::Signer, transaction::Transaction};
 use spl_associated_token_account::get_associated_token_address;
@@ -12,9 +13,12 @@ use stakedex_sdk::Stakedex;
 use stakedex_sdk_common::{bsol, esol, jitosol, jsol, msol, stsol};
 use std::{collections::HashMap, iter::zip, str::FromStr};
 
-const WHALE: &str = "9uyDy9VDBw4K7xoSkhmCAm8NAFCwu4pkF6JeHUCtVKcX";
+// Alameda account with yuge mSOL, jSOL, stSOL holdings
+pub mod whale {
+    solana_sdk::declare_id!("9uyDy9VDBw4K7xoSkhmCAm8NAFCwu4pkF6JeHUCtVKcX");
+}
 
-mod jupiter_program {
+pub mod jupiter_program {
     // NOT IN USE, JUST BECAUSE ITS REQUIRED AS A STRUCT FIELD FOR jupiter_amm_interface::SwapParams
     solana_sdk::declare_id!("JUP4Fb2cqiRUcaTHdrPC8h2gNsA2ETXiPDD33WcGuJB");
 }
@@ -51,7 +55,7 @@ fn fetch_accounts(accounts_pubkeys: &[Pubkey]) -> HashMap<Pubkey, Account> {
 }
 
 #[test]
-fn test_swap_via_stake_jitosol_bsol() {
+fn test_quote_swap_via_stake_jitosol_bsol() {
     STAKEDEX
         .quote_swap_via_stake(&QuoteParams {
             in_amount: 1_000_000_000,
@@ -62,7 +66,7 @@ fn test_swap_via_stake_jitosol_bsol() {
 }
 
 #[test]
-fn test_swap_via_stake_esol_bsol() {
+fn test_quote_swap_via_stake_esol_bsol() {
     STAKEDEX
         .quote_swap_via_stake(&QuoteParams {
             in_amount: 1_000_000_000, // 1_000_000_000_000
@@ -73,7 +77,7 @@ fn test_swap_via_stake_esol_bsol() {
 }
 
 #[test]
-fn test_swap_via_stake_unknown_token() {
+fn test_quote_swap_via_stake_unknown_token() {
     let unknown_token = Pubkey::new_unique();
     let res = STAKEDEX.quote_swap_via_stake(&QuoteParams {
         in_amount: 1_000_000_000,
@@ -83,71 +87,56 @@ fn test_swap_via_stake_unknown_token() {
     assert!(res.is_err());
 }
 
-fn test_swap_via_stake(input_mint: Pubkey, output_mint: Pubkey) {
-    let whale_pk = Pubkey::from_str(WHALE).unwrap();
-    let res = STAKEDEX.quote_swap_via_stake(&QuoteParams {
-        in_amount: 100_000_000_000,
-        input_mint,
-        output_mint,
-    });
-    match res {
-        Err(err) => assert!(err.to_string() == "No route found between pools"),
-        Ok(quote) => {
-            println!("{:?}", quote);
-            let source_token_account = get_associated_token_address(&whale_pk, &input_mint);
-            let destination_token_account = get_associated_token_address(&whale_pk, &output_mint);
-            let ix = STAKEDEX
-                .swap_via_stake_ix(
-                    &SwapParams {
-                        jupiter_program_id: &jupiter_program::ID,
-                        in_amount: quote.in_amount,
-                        destination_mint: output_mint,
-                        source_mint: input_mint,
-                        user_destination_token_account: destination_token_account,
-                        user_source_token_account: source_token_account,
-                        user_transfer_authority: whale_pk,
-                        open_order_address: None,
-                        quote_mint_to_referrer: None,
-                    },
-                    0,
-                )
-                .unwrap();
-            // let msg = Message::new(&[ix], Some(&whale_pk));
-            // let blockhash = RPC.get_latest_blockhash().unwrap();
-            let tx = Transaction::new_with_payer(&[ix], Some(&whale_pk));
-            let result = RPC
-                .simulate_transaction_with_config(
-                    &tx,
-                    RpcSimulateTransactionConfig {
-                        accounts: Some(RpcSimulateTransactionAccountsConfig {
-                            addresses: vec![
-                                source_token_account.to_string(),
-                                destination_token_account.to_string(),
-                            ],
-                            encoding: Some(UiAccountEncoding::Base64),
-                        }),
-                        ..RpcSimulateTransactionConfig::default()
-                    },
-                )
-                .unwrap();
-            println!(
-                "1 {}\n2 {}",
-                source_token_account, destination_token_account
-            );
-            println!("Simulate result: {:?}", result.value);
-            println!("{:?}", result.value.accounts.unwrap());
-        }
-    }
-}
-
 #[test]
 fn test_swap_via_stake_stsol_unstakeit() {
-    test_swap_via_stake(stsol::ID, native_mint::ID);
+    let resp = sim_swap_via_stake_with_whale(&STAKEDEX, &RPC, stsol::ID, native_mint::ID);
+    assert_sim_success(&resp);
 }
 
 #[test]
 fn test_swap_via_stake_jsol_msol() {
-    test_swap_via_stake(jsol::ID, msol::ID);
+    let resp = sim_swap_via_stake_with_whale(&STAKEDEX, &RPC, jsol::ID, msol::ID);
+    assert_sim_success(&resp);
+}
+
+#[test]
+fn test_swap_via_stake_jitosol_bsol() {
+    let signer = Pubkey::from_str("71WDyyCsZwyEYDV91Qrb212rdg6woCHYQhFnmZUBxiJ6").unwrap();
+    let src_token_acc = Pubkey::from_str("83ohMPRdV5XJ868EmbDHyKEgK1gdaSctHpiXVcWiPBn7").unwrap();
+    let dst_token_acc = Pubkey::from_str("9D1JFhFFd4rHWdJfjVkBaD6u3ZKcGtz7W4N1mqYUaM3T").unwrap();
+    let resp = sim_swap_via_stake(
+        &STAKEDEX,
+        &RPC,
+        TestSwapViaStakeArgs {
+            in_amount: 100_000_000,
+            input_mint: jitosol::ID,
+            output_mint: bsol::ID,
+            signer,
+            src_token_acc,
+            dst_token_acc,
+        },
+    );
+    assert_sim_success(&resp);
+}
+
+#[test]
+fn test_swap_via_stake_bsol_jitosol() {
+    let signer = Pubkey::from_str("71WDyyCsZwyEYDV91Qrb212rdg6woCHYQhFnmZUBxiJ6").unwrap();
+    let src_token_acc = Pubkey::from_str("9D1JFhFFd4rHWdJfjVkBaD6u3ZKcGtz7W4N1mqYUaM3T").unwrap();
+    let dst_token_acc = Pubkey::from_str("83ohMPRdV5XJ868EmbDHyKEgK1gdaSctHpiXVcWiPBn7").unwrap();
+    let resp = sim_swap_via_stake(
+        &STAKEDEX,
+        &RPC,
+        TestSwapViaStakeArgs {
+            in_amount: 100_000_000,
+            input_mint: bsol::ID,
+            output_mint: jitosol::ID,
+            signer,
+            src_token_acc,
+            dst_token_acc,
+        },
+    );
+    assert_sim_success(&resp);
 }
 
 #[test]
@@ -196,26 +185,114 @@ fn test_jsol_drain_vsa_edge_case() {
     });
     assert!(should_fail.is_err());
     // try simulating max possible quote
-    let whale_pk = Pubkey::from_str(WHALE).unwrap();
-    let user_source_token_account = get_associated_token_address(&whale_pk, &jsol::ID);
-    let user_destination_token_account = get_associated_token_address(&whale_pk, &msol::ID);
-    let params = SwapParams {
-        jupiter_program_id: &jupiter_program::ID,
-        in_amount: max_possible_quote.in_amount,
-        destination_mint: msol::ID,
-        source_mint: jsol::ID,
-        user_destination_token_account,
-        user_source_token_account,
-        user_transfer_authority: whale_pk,
-        open_order_address: None,
-        quote_mint_to_referrer: None,
-    };
-    let ix = STAKEDEX.swap_via_stake_ix(&params, 0).unwrap();
-    let mut tx = Transaction::new_with_payer(&[ix], Some(&whale_pk));
-    let rbh = RPC.get_latest_blockhash().unwrap();
+    let result = sim_swap_via_stake(
+        &STAKEDEX,
+        &RPC,
+        TestSwapViaStakeArgs {
+            in_amount: max_possible_quote.in_amount,
+            input_mint: jsol::ID,
+            output_mint: msol::ID,
+            signer: whale::ID,
+            src_token_acc: get_associated_token_address(&whale::ID, &jsol::ID),
+            dst_token_acc: get_associated_token_address(&whale::ID, &msol::ID),
+        },
+    );
+    assert!(result.value.err.is_none());
+}
+
+pub struct TestSwapViaStakeArgs {
+    pub in_amount: u64,
+    pub input_mint: Pubkey,
+    pub output_mint: Pubkey,
+    pub signer: Pubkey,
+    pub src_token_acc: Pubkey,
+    pub dst_token_acc: Pubkey,
+}
+
+pub fn sim_swap_via_stake_with_whale(
+    stakedex: &Stakedex,
+    rpc: &RpcClient,
+    input_mint: Pubkey,
+    output_mint: Pubkey,
+) -> Response<RpcSimulateTransactionResult> {
+    sim_swap_via_stake(
+        stakedex,
+        rpc,
+        TestSwapViaStakeArgs {
+            in_amount: 100_000_000_000,
+            input_mint,
+            output_mint,
+            signer: whale::ID,
+            src_token_acc: get_associated_token_address(&whale::ID, &input_mint),
+            dst_token_acc: get_associated_token_address(&whale::ID, &output_mint),
+        },
+    )
+}
+
+fn assert_sim_success(response: &Response<RpcSimulateTransactionResult>) {
+    if let Some(e) = &response.value.err {
+        eprintln!("{:#?}", response.value.logs.as_ref().unwrap());
+        eprintln!("ERROR: {e}");
+    }
+}
+
+pub fn sim_swap_via_stake(
+    stakedex: &Stakedex,
+    rpc: &RpcClient,
+    TestSwapViaStakeArgs {
+        in_amount,
+        input_mint,
+        output_mint,
+        signer,
+        src_token_acc,
+        dst_token_acc,
+    }: TestSwapViaStakeArgs,
+) -> Response<RpcSimulateTransactionResult> {
+    let quote = stakedex
+        .quote_swap_via_stake(&QuoteParams {
+            in_amount,
+            input_mint,
+            output_mint,
+        })
+        .unwrap();
+    // println!("{:?}", quote);
+    let ix = stakedex
+        .swap_via_stake_ix(
+            &SwapParams {
+                jupiter_program_id: &jupiter_program::ID,
+                in_amount: quote.in_amount,
+                destination_mint: output_mint,
+                source_mint: input_mint,
+                user_destination_token_account: dst_token_acc,
+                user_source_token_account: src_token_acc,
+                user_transfer_authority: signer,
+                open_order_address: None,
+                quote_mint_to_referrer: None,
+            },
+            0,
+        )
+        .unwrap();
+    // let msg = Message::new(&[ix], Some(&whale_pk));
+    // let blockhash = RPC.get_latest_blockhash().unwrap();
+    let rbh = rpc.get_latest_blockhash().unwrap();
+    let mut tx = Transaction::new_with_payer(&[ix], Some(&signer));
     // partial_sign just to add recentblockhash
     let no_signers: Vec<Box<dyn Signer>> = vec![];
     tx.partial_sign(&no_signers, rbh);
-    let result = RPC.simulate_transaction(&tx).unwrap();
-    assert!(result.value.err.is_none());
+    let result = rpc
+        .simulate_transaction_with_config(
+            &tx,
+            RpcSimulateTransactionConfig {
+                accounts: Some(RpcSimulateTransactionAccountsConfig {
+                    addresses: vec![src_token_acc.to_string(), dst_token_acc.to_string()],
+                    encoding: Some(UiAccountEncoding::Base64),
+                }),
+                ..RpcSimulateTransactionConfig::default()
+            },
+        )
+        .unwrap();
+    if let Some(err) = result.value.err {
+        panic!("{err}");
+    }
+    result
 }
